Use libreadline-gplv2-dev instead of libreadline-dev
Upstream LCNC uses `libreadline-gplv2-dev`, the GPLv2-compatible
readline library package.  MK still uses the conflicting
`libreadline-dev` package which is not GPLv2-compatible.  That has
been fixed in this branch, but not upstream.  This patch pulls in a
change from my private MK-CNC tree to resolve this on the MK-CNC side.